Refactor the way to get the pokemon data from the api and how to save it in the store
