+++ src/components/styles/pokedexitem.module.css
@@ -11,7 +11,7 @@
 }
 
 .Item:hover {
-  background-color: var(--color-5);
+  background-color: rgb(0, 0, 0, 0.3);
   cursor: pointer;
 }
