+++ src/index.js
@@ -6,7 +6,7 @@ import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import Routes from './routes';
 import rootReducer from './reducers/root';
-import * as ApiComms from './apicomms/apicomms';
+import * as ApiComms from './modules/apicomms';
 import * as Actions from './actions/actions';
 import AppStyles from './components/styles/app.module.css';
 
@@ -15,9 +15,35 @@ const store = createStore(
   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__(),
 );
 
+const getPokemonsInStore = async pokemon => {
+  const pokemonInfo = await ApiComms.getOnePokemonInfo(pokemon.name);
+  const {
+    id,
+    name,
+    sprites: { front_default: imagesm },
+    sprites: { other: { dream_world: { front_default: imagelg } } },
+    types,
+    weight,
+    height,
+  } = pokemonInfo;
+
+  const onePokemon = {
+    id,
+    name,
+    imagesm,
+    imagelg,
+    types,
+    weight,
+    height,
+  };
+  store.dispatch(Actions.addPokemon(onePokemon));
+};
+
 const dataInit = async () => {
   const { results: pokemons } = await ApiComms.getPokemons(1);
-  store.dispatch(Actions.addPokemon(pokemons));
+  pokemons.forEach(pokemon => {
+    getPokemonsInStore(pokemon);
+  });
 };
 
 dataInit();
