+++ src/reducers/pokedex.js
@@ -4,9 +4,7 @@ const pokedexReducer = (state = defaultstate, action) => {
   const { type, payload: pokemon } = action;
   switch (type) {
     case 'ADD_POKEMON':
-      return pokemon;
-    case 'CHANGE_POKEMON':
-      return state;
+      return state.concat(pokemon);
     default:
       return state;
   }
